Refactor code and protocol with assistant correction
